+++ include/cipher.bip32.path.go.h
@@ -1,9 +1,9 @@
 typedef struct {
-    GoSlice_ Elements;
+    GoSlice_ Elements; // Goslice de PathNote
 } Path;
 
 // PathNode is an element of an HD wallet path
-typedef struct PathNode {
-    bool Master;
+typedef struct {
+    BOOL Master;
     GoUint32_ ChildNumber;
-}
+} PathNode;